optimization: [internal-s13] a collection of modifications according audit feedback

+++ src/libraries/ProtocolFeeLibrary.sol
@@ -5,6 +5,7 @@ pragma solidity ^0.8.0;
 import "./math/UnsafeMath.sol";
 
 library ProtocolFeeLibrary {
+    /// @dev Increasing these values could lead to overflow in Pool.swap
     /// @notice Max protocol fee is 0.1% (1000 pips)
     uint16 public constant MAX_PROTOCOL_FEE = 1000;
 
@@ -44,7 +45,10 @@ library ProtocolFeeLibrary {
     /// @notice The protocol fee is taken from the input amount first and then the LP fee is taken from the remaining
     /// Equivalent to `protocolFee + lpFee(1_000_000 - protocolFee) / 1_000_000`
     /// Hence the swap fee is capped at 1_000_000 (100%) for cl pool and 100_000 (10%) for bin pool
-    function calculateSwapFee(uint24 self, uint24 lpFee) internal pure returns (uint24 swapFee) {
+    /// @param self The single direction protocol fee to calculate the swap fee from
+    /// @param lpFee The LP fee to calculate the swap fee from
+    /// @return swapFee The composite swap fee
+    function calculateSwapFee(uint16 self, uint24 lpFee) internal pure returns (uint24 swapFee) {
         assembly ("memory-safe") {
             let numerator := mul(self, lpFee)
             let divRoundingUp := add(div(numerator, PIPS_DENOMINATOR), gt(mod(numerator, PIPS_DENOMINATOR), 0))
